+++ python/src/runapi/suno/resources/inspire_music.py
@@ -1,0 +1,9 @@
+from .audio_actions import AudioAction
+
+
+class InspireMusic(AudioAction):
+    ENDPOINT = "/api/v1/suno/inspire_music"
+    ACTION = "inspire-music"
+
+    def create(self, options=None, **params):
+        return super().create(options=options, **params)